requirements.txt
discord.py
ffmpeg
pytest
yt-dlp
PyNaCl
youtube_dl